chore: Update dependencies and import statements
This commit updates the dependencies in the `package.json` file and the `yarn.lock` file. It also updates the import statements in the Dash pages to remove the unnecessary "lang" attribute. These changes improve code readability, consistency, and ensure that the project is using the latest versions of the dependencies.